+++ action.dispatch.event_test.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/dogmatiq/configkit"
-	"github.com/dogmatiq/configkit/message"
 	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/enginekit/enginetest/stubs"
 	. "github.com/dogmatiq/testkit"
@@ -75,8 +74,6 @@ var _ = g.Describe("func RecordEvent()", func() {
 					CausationID:   "1",
 					CorrelationID: "1",
 					Message:       EventA1,
-					Type:          message.TypeOf(EventA1),
-					Role:          message.EventRole,
 					CreatedAt:     startTime,
 				},
 				EngineTime: startTime,
@@ -101,19 +98,6 @@ var _ = g.Describe("func RecordEvent()", func() {
 		gm.Expect(t.Failed()).To(gm.BeTrue())
 		gm.Expect(t.Logs).To(gm.ContainElement(
 			"cannot record event, stubs.EventStub[TypeX] is a not a recognized message type",
-		))
-	})
-
-	g.It("fails the test if the message type is not an event", func() {
-		t.FailSilently = true
-
-		test.Prepare(
-			RecordEvent(CommandA1),
-		)
-
-		gm.Expect(t.Failed()).To(gm.BeTrue())
-		gm.Expect(t.Logs).To(gm.ContainElement(
-			"cannot record event, stubs.CommandStub[TypeA] is configured as a command",
 		))
 	})