+++ engine/configurer.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/dogmatiq/configkit"
-	"github.com/dogmatiq/configkit/message"
 	"github.com/dogmatiq/testkit/engine/internal/aggregate"
 	"github.com/dogmatiq/testkit/engine/internal/integration"
 	"github.com/dogmatiq/testkit/engine/internal/process"
@@ -17,70 +16,55 @@ type configurer struct {
 }
 
 func (c *configurer) VisitRichApplication(ctx context.Context, cfg configkit.RichApplication) error {
-	c.engine.roles = cfg.MessageTypes().All()
-
 	return cfg.RichHandlers().AcceptRichVisitor(ctx, c)
 }
 
 func (c *configurer) VisitRichAggregate(_ context.Context, cfg configkit.RichAggregate) error {
-	mt := cfg.MessageTypes()
 	c.registerController(
 		&aggregate.Controller{
 			Config:     cfg,
 			MessageIDs: &c.engine.messageIDs,
 		},
-		mt.Consumed,
 	)
-
 	return nil
 }
 
 func (c *configurer) VisitRichProcess(_ context.Context, cfg configkit.RichProcess) error {
-	mt := cfg.MessageTypes()
 	c.registerController(
 		&process.Controller{
 			Config:     cfg,
 			MessageIDs: &c.engine.messageIDs,
 		},
-		mt.Consumed,
 	)
-
 	return nil
 }
 
 func (c *configurer) VisitRichIntegration(_ context.Context, cfg configkit.RichIntegration) error {
-	mt := cfg.MessageTypes()
 	c.registerController(
 		&integration.Controller{
 			Config:     cfg,
 			MessageIDs: &c.engine.messageIDs,
 		},
-		mt.Consumed,
 	)
-
 	return nil
 }
 
 func (c *configurer) VisitRichProjection(_ context.Context, cfg configkit.RichProjection) error {
-	mt := cfg.MessageTypes()
 	c.registerController(
 		&projection.Controller{
 			Config:                cfg,
 			CompactDuringHandling: c.options.compactDuringHandling,
 		},
-		mt.Consumed,
 	)
-
 	return nil
 }
 
-func (c *configurer) registerController(
-	ctrl controller,
-	types map[message.Type]message.Role,
-) {
-	c.engine.controllers[ctrl.HandlerConfig().Identity().Name] = ctrl
+func (c *configurer) registerController(ctrl controller) {
+	cfg := ctrl.HandlerConfig()
 
-	for t := range types {
+	c.engine.controllers[cfg.Identity().Name] = ctrl
+
+	for t := range cfg.MessageTypes().Consumed() {
 		c.engine.routes[t] = append(c.engine.routes[t], ctrl)
 	}
 }