+++ internal/inflect/inflect.go
@@ -9,8 +9,8 @@ import (
 	"github.com/dogmatiq/configkit/message"
 )
 
-var substitutions = map[message.Role]map[string]string{
-	message.CommandRole: {
+var substitutions = map[message.Kind]map[string]string{
+	message.CommandKind: {
 		"<message>":    "command",
 		"<messages>":   "commands",
 		"<produce>":    "execute",
@@ -18,7 +18,7 @@ var substitutions = map[message.Role]map[string]string{
 		"<producing>":  "executing",
 		"<dispatcher>": "dogma.CommandExecutor",
 	},
-	message.EventRole: {
+	message.EventKind: {
 		"<message>":    "event",
 		"<messages>":   "events",
 		"<produce>":    "record",
@@ -26,7 +26,7 @@ var substitutions = map[message.Role]map[string]string{
 		"<producing>":  "recording",
 		"<dispatcher>": "dogma.EventRecorder",
 	},
-	message.TimeoutRole: {
+	message.TimeoutKind: {
 		"<message>":   "timeout",
 		"<messages>":  "timeouts",
 		"<produce>":   "schedule",
@@ -44,9 +44,9 @@ var corrections = map[string]string{
 	"1 timeouts": "1 timeout",
 }
 
-// Sprint formats a string, inflecting words in s match the message role r.
-func Sprint(r message.Role, s string) string {
-	for k, v := range substitutions[r] {
+// Sprint formats a string, inflecting words in s match the message kind k.
+func Sprint(k message.Kind, s string) string {
+	for k, v := range substitutions[k] {
 		s = strings.ReplaceAll(s, k, v)
 		s = strings.ReplaceAll(s, strings.ToUpper(k), strings.ToUpper(v))
 	}
@@ -59,22 +59,22 @@ func Sprint(r message.Role, s string) string {
 	return s
 }
 
-// Sprintf formats a string, inflecting words in f match the message role r.
-func Sprintf(r message.Role, f string, v ...any) string {
+// Sprintf formats a string, inflecting words in f match the message kind k.
+func Sprintf(k message.Kind, f string, v ...any) string {
 	return Sprint(
-		r,
+		k,
 		fmt.Sprintf(f, v...),
 	)
 }
 
-// Error returns a new error, inflecting words in s to match the message role r.
-func Error(r message.Role, s string) error {
-	return errors.New(Sprint(r, s))
+// Error returns a new error, inflecting words in s to match the message kind k.
+func Error(k message.Kind, s string) error {
+	return errors.New(Sprint(k, s))
 }
 
-// Errorf returns a new error, inflecting words in f to match the message role r.
-func Errorf(r message.Role, f string, v ...any) error {
-	return errors.New(Sprintf(r, f, v...))
+// Errorf returns a new error, inflecting words in f to match the message kind k.
+func Errorf(k message.Kind, f string, v ...any) error {
+	return errors.New(Sprintf(k, f, v...))
 }
 
 // replaceAll replaces all instances of k with v, at word boundaries.