+++ internal/validation/doc.go
@@ -1,0 +1,2 @@
+// Package validation provides utilities for validation of Dogma message types.
+package validation